bug found!!! wheel radius 0.03!!!!!!

+++ 0527_peanuts/Core/Inc/mainpp.h
@@ -17,7 +17,7 @@ void setup(void);
 void loop(void);
 void errcallback(void);
 void realspeed(void);
-extern float Vx, Vy, W;
+extern double Vx, Vy, W;
 extern float rVx, rVy, rW;
 
 #ifdef __cplusplus
